BREAKING: remove eval-in-project
Leiningen version 2.8.0 and later no longer ships its own version of Lucene,
which means it's no longer necessary to run dependency-check in the project
context. Moreover, it's undesirable to run in the project context due to the
potential for conflicts with the project's own dependencies (including Lucene!).

With this change, the `nvd-clojure` module becomes a static dependency of the plugin.
This breaks the ability to dynamically load the latest `nvd-clojure` version
at runtime, as well as the abiliy to pin the version used (though version pinning
would no longer be necessary).