Fix handling of managed dependencies edge cases
The existing logic for filtering the dependency hierarchy misses some
edge cases, including the case where a dependency is declared without a
version (e.g. `:dependencies [[foo]]`) and the corresponding managed
dependency is declared in another `project.clj` file (as can be the case
when using plugins like lein-parent).

This change addresses these edge cases by delegating dependency
resolution to Leiningen.

Based on the commit history it appears that `leiningen.nvd.deps` was
was added to filter the classpath to only include dependencies declared
in a `:dependencies` or `:managed-dependencies` list. In particular,
there was a desire to exclude `:plugins` from the scanned classpath.
Based on #1 I believe there was also a desire to exclude certain base
profiles (presumably `:user` and `:base`).

Current versions of Leiningen do not appear to include `:plugins` in
`get-classpath` computation, but do include some other directories which
are irrelevant in this case (e.g. the local src and resource
directories). Rather than reverting back to using `get-classpath` this
change uses `resolve-managed-dependencies` to exclude these extraneous
classpath elements. The `:base`, `:system` and `:user` profiles are also
unmerged so that only dependencies of the project itself are included.