Derive the classpath for Clojure CLI from deps.edn
The classpath generation in Clojure CLI differs to classpath
generation with Leiningen: with Clojure CLI even the JARs needed for
lein-nvd would be picked up and added to the classpath. Which isn't
what you want when checking your own project/library. To work around
that, the classpath for Clojure CLI is built with just the artifacts
specified in deps.edn.

The assumption is made that if a library/project contains deps.edn,
only that will be read to work the classpath. I'm mentioning this in
the context of projects which have both, deps.edn and project.clj
included in them.