Added tests for classpath
The functions that generate classpaths spit out strings that are full
paths to the JARs on classpath.

I'm just checking if Clojure jar is present.

In the process I've added deps.edn since it's really hard to change
the current working directory, just for tests. But having the deps.edn
allows people to use this library via git sha, via Clojure CLI, rather
than just the versions pushed to Clojars.