Automated dependency management (#67)
I've taken the liberty to implement automated dependency management.

Why?

The badge pointing to dependency versioning service, does not work.

But even if it did, wouldn't it be more convenient to have something
that let's you know that dependencies not up to date? And what more
convenient than an automated CI job that does just that?

I took the liberty of creating just that workflow. It does the
following:

- it checks for out of date dependencies
- if there are any, updates the badge
- but it also creates the PR with the updated dependencies

And since I've added deps.edn and the project uses Leiningen, I
thought that it would be convenient to automate the updating of both
deps.edn and project.clj with the outdated dependencies.