Update dependencies in project.clj (#71)
outdated.md needs to be removed before creating the PR otherwise it
makes it into the PR.

:no-tests directive needs to be used with lein ancient, so that the
test aren't run straight away, after the dependency upgrade. This way
the PR is created.

Whoever gets the PR will still need to do some manual clean up, but at
least the start of the PR is there and it's visible that the upgrade
should take place.