Adjust dependencies
This removes explicit dependencies that are no longer needed, either
because newer versions are brought in transitively, or, in the case of
maven-core, maven-resolver-transport-http, and jetty-client, are no
longer used.

This also upgrades sljf4j-simple to satisfy pedantic, as
dependency-check-core brings in two versions transitively, but 2.0.12 is
the greater of the two.